refactor: organize optional features into features/ module
Move compression, hashing, and detection modules into a dedicated
features/ subdirectory to better separate optional functionality from
core encoding/decoding operations.

Changes:
- Create src/features/ directory structure
- Move compression.rs, hashing.rs, detection.rs to features/
- Add features/mod.rs with re-exports for backward compatibility
- Update src/lib.rs to use features module
- Update imports in streaming encoder/decoder to use crate::features::

All public APIs remain unchanged - this is purely an internal
reorganization. Tests pass successfully (229/229).
