+++ src/features/hashing.rs
